refactor: move to other naming convention of filenames
